refactor(i18n): rename thread page i18n file for clarity
- Renamed `i18n.tsx` to `thread-page-i18n.tsx` to improve file naming consistency
- Updated import path in `i18n/config.ts` to reflect the new file name
