+++ ts-packages/web/src/app/(social)/threads/[id]/thread-page-i18n.tsx
